backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.3
cryptography==45.0.3
cycler==0.12.1
dateparser==1.2.2
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fake-useragent==2.2.0
fastapi==0.115.0
feedparser==6.0.11
flake8==7.1.1
fonttools==4.59.1
frozenlist==1.7.0
fuzzywuzzy==0.18.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
jiter==0.10.0
joblib==1.5.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Levenshtein==0.27.1
markdown-it-py==4.0.0
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
nltk==3.9.1
numpy==2.2.6
openai==1.105.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.11
PyAudio==0.2.14
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypiwin32==223
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.37.0
regex==2025.9.1
requests==2.32.3
rich==14.1.0
rpds-py==0.28.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.2.0
selenium-stealth==1.0.6
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.38.6
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
urllib3-secure-extra==0.1.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
whitenoise==6.7.0
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.20.1
